+++ src/components/TodoComponents/Todo.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './Todo.css';
+
+const Todo = props => {
+    const { task } = props.toDoProp
+    return <div>{task}</div>;
+};
+
+
+export default Todo;
