+++ src/components/TodoComponents/TodoList.js
@@ -1,2 +1,16 @@
 // your components will all go in this `component` directory.
 // feel free to change this component.js into TodoList.js
+
+import React from 'react';
+import './Todo.css';
+import Todo from './Todo';
+
+const TodoList = props => {
+    return <div>{props.toDoListProp.map(todo => <Todo key={todo.id} toDoProp={todo} />)}</div>;
+};
+
+
+
+export default TodoList;
+
+
